requirements.txt
openai
langchain
langchain-openai
python-dotenv
langchain_community
faiss-cpu
streamlit
langchain-huggingface